+++ Changelog.md
@@ -3,7 +3,8 @@
 ## [1.3.1] - 25-04-2025
 ### Added
 - Get appearance settings from admin panel to override the default attributes but not the ones from init.
-- Make locale required 
+- Make locale required
+- Add event listener to close the popup
 
 ## [1.2.2] - 02-04-2025
 ### Added